Rename tests/utils.py into tests/_validation.py
