+++ tests/test_workflows.py
@@ -21,8 +21,8 @@ from typing import List
 import pytest
 from devtools import debug
 
-from .utils import check_file_number
-from .utils import validate_schema
+from ._validation import check_file_number
+from ._validation import validate_schema
 from fractal_tasks_core.tasks.copy_ome_zarr import copy_ome_zarr
 from fractal_tasks_core.tasks.create_ome_zarr import create_ome_zarr
 from fractal_tasks_core.tasks.illumination_correction import (
