+++ tests/test_workflows_cellpose_segmentation.py
@@ -30,10 +30,10 @@ from devtools import debug
 from pytest import MonkeyPatch
 
 import fractal_tasks_core.tasks
+from ._validation import check_file_number
+from ._validation import validate_axes_and_coordinateTransformations
+from ._validation import validate_schema
 from .lib_empty_ROI_table import _add_empty_ROI_table
-from .utils import check_file_number
-from .utils import validate_axes_and_coordinateTransformations
-from .utils import validate_schema
 from fractal_tasks_core.tasks.cellpose_segmentation import (
     cellpose_segmentation,
 )
