+++ tests/test_workflows_napari_workflows.py
@@ -22,11 +22,11 @@ import anndata as ad
 import pytest
 from devtools import debug
 
+from ._validation import check_file_number
+from ._validation import validate_axes_and_coordinateTransformations
+from ._validation import validate_labels_and_measurements
+from ._validation import validate_schema
 from .lib_empty_ROI_table import _add_empty_ROI_table
-from .utils import check_file_number
-from .utils import validate_axes_and_coordinateTransformations
-from .utils import validate_labels_and_measurements
-from .utils import validate_schema
 from fractal_tasks_core.tasks.napari_workflows_wrapper import (
     napari_workflows_wrapper,
 )
